+++ scripts/trim-low-abund.py
@@ -51,14 +51,14 @@ import khmer
 import tempfile
 import shutil
 import textwrap
-import argparse
 
 from khmer import khmer_args
 from khmer import ReadParser
 
-from khmer.khmer_args import (build_counting_args, info, add_loadgraph_args,
+from khmer.khmer_args import (build_counting_args, add_loadgraph_args,
                               report_on_config, calculate_graphsize,
                               sanitize_help)
+from khmer.khmer_args import FileType as khFileType
 from khmer.utils import (write_record, write_record_pair, broken_paired_reader,
                          ReadBundle)
 from khmer.kfile import (check_space, check_space_for_graph,
@@ -100,7 +100,8 @@ def get_parser():
 
     parser = build_counting_args(
         descr='Trim low-abundance k-mers using a streaming algorithm.',
-        epilog=textwrap.dedent(epilog))
+        epilog=textwrap.dedent(epilog),
+        citations=['streaming'])
 
     parser.add_argument('input_filenames', nargs='+')
 
@@ -114,7 +115,7 @@ def get_parser():
                         default=DEFAULT_TRIM_AT_COVERAGE)
 
     parser.add_argument('-o', '--output', metavar="output_filename",
-                        type=argparse.FileType('wb'),
+                        type=khFileType('wb'),
                         help='only output a single file with '
                         'the specified filename; use a single dash "-" to '
                         'specify that output should go to STDOUT (the '
@@ -274,8 +275,6 @@ class Trimmer(object):
 def main():
     parser = sanitize_help(get_parser())
     args = parser.parse_args()
-    if not args.quiet:
-        info('trim-low-abund.py', ['streaming'])
 
     configure_logging(args.quiet)